Move checkout shipping commands to checkout folder
