+++ web/app/integration-manager/commands.js
@@ -3,6 +3,7 @@ import {register as registerProducts} from './products/commands'
 import {register as registerCategories} from './categories/commands'
 import {register as registerCart} from './cart/commands'
 import {register as registerApp} from './app/commands'
+import {register as registerCheckout} from './checkout/commands'
 
 let connector = {}
 
@@ -13,16 +14,9 @@ export const register = (commands) => {
     registerProducts(commands.products)
     registerCategories(commands.categories)
     registerCart(commands.cart)
+    registerCheckout(commands.checkout)
 }
 
 export const addToCart = (...args) => connector.addToCart(...args)
-
-export const fetchCheckoutShippingData = (...args) => connector.fetchCheckoutShippingData(...args)
-
-export const submitShipping = (...args) => connector.submitShipping(...args)
-
-export const checkCustomerEmail = (...args) => connector.checkCustomerEmail(...args)
-
-export const submitSignIn = (...args) => connector.submitSignIn(...args)
 
 export const submitNewsletter = (...args) => connector.submitNewsletter(...args)
