+++ web/app/integration-manager/responses.js
@@ -1,8 +1,6 @@
 import {createAction} from '../utils/utils'
 import {addNotification, removeNotification} from '../containers/app/actions'
 
-export const receiveCheckoutShippingData = createAction('Receive Checkout Shipping data')
-export const receiveCheckoutData = createAction('Receive Checkout Data')
 export const receiveHomeData = createAction('Receive Home Data')
 export const receiveNavigationData = createAction('Receive Navigation Data')
 export const receiveFooterData = createAction('Receive Footer Data')
